chore: modify contentlayer config
