--- contentlayer.config.ts
@@ -1,18 +1,0 @@
-import { defineDocumentType, makeSource } from 'contentlayer/source-files';
-
-export const Post = defineDocumentType(() => ({
-  name: 'Posts',
-  filePathPattern: `**/*.mdx`,
-  fields: {
-    title: { type: 'string', required: true },
-    date: { type: 'date', required: true },
-  },
-  computedFields: {
-    slug: {
-      type: 'string',
-      resolve: (post) => `${post._raw.flattenedPath}`,
-    },
-  },
-}));
-
-export default makeSource({ contentDirPath: 'posts', documentTypes: [Post] });
